+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 * Host pool election for compute allocation can be more relevant ([GH-83](https://github.com/ystia/yorc/issues/83))
 
+### BUG FIXES
+
+* Tosca public_ip_address attribute is wrongly set with private address for hosts pool computes ([GH-593](https://github.com/ystia/yorc/issues/593))
+
 ## 4.0.0-M8 (January 24, 2020)
 
 ### BREAKING CHANGES